requirements.txt

asgiref
certifi
crispy-bootstrap5
distlib
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-email
django-environ
django-filter
django-humanize
djangorestframework
EasyProcess
entrypoint2
filelock
humanize
Markdown
MouseInfo
mss
mysql-connector-python
Pillow
pipenv
platformdirs
protobuf
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
pyscreenshot
PyScreeze
pytweening
sqlparse
typing_extensions
tzdata
virtualenv